langchain/requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
langchain-text-splitters
fastapi
uvicorn
python-dotenv
pydantic
python-multipart
faiss-cpu
tiktoken
openai 